Implement the Miller-Rabin test using ints
I created a copy of the Miller-Rabin code that uses unsigned longs
instead of mpz_t to hold the numbers.  This limits the test to 32 bit
numbers. (Larger numbes will, when multiplied, produce results to large
to fit into a register, so the results will be cut off, making their
reduction modulo n plain wrong.)

This implementation can be used to run the Miller-Rabin test on all odd
numbers in the intervall [5, 2^32].  It is mainly intended to compare
the runtime and error rate to the RQFT.