Refactor code in frobenius_int.c and fix int_sqrt
Change the int_sqrt implementation to work for number that are to large
to get a good estimate of the integer square root by converting to
double, taking the square root and truncating the result.  The new
implementation is not particularly good, as the floating point
approximation is used as a starting point, from which *every* integer
is tested until one is found which is larger than the square root.  The
one before that will then be returned.

Remove all "_long"-suffixes, as the GMP implementation is the main one.
Move the code handling the first two steps into a new function which
only takes the integer square root once instead of twice and reduces
some code duplication.