A chain addition takes just 4 multiplications
The original assumption of 8 multiplications is for computing more general
products.  Here, we have B_1 = 1, so we can avoid a few multiplications.
In addition, k/2 mod n can be computed by shifting to the right (if k is
even) or shifting (k+n) to the right (if k is odd).