Add small programmes to check the RQFT on small inputs
There are two programmes, each in a long long and a GMP version.  The
programme "check_all_params" computes the QFT of all composite 5<=n<10000
and counts the number of non-witnesses.  "check_all_small_numbers" chooses
the parameters (b,c) mod n in a simple deterministic fashion and looks for
false positives.